Source/Server/requirements.txt
click>=5.1
itsdangerous>=0.24
jinja2>=2.10.1
werkzeug>=0.15
Flask==1.1.2
pyotp==2.6.0
Flask-SQLAlchemy==2.5.1
PyJWT==2.4.0
Faker==13.14.0
faker-vehicle==0.2.0
SQLAlchemy-Utils==0.38.2
cryptography==37.0.3